requirements.txt
cadquery-ocp==7.8.1.1.post1
dotenv
openai
pydantic
